front: refactorization: add modules => scenario, study, project, rollingStock, conflict
