+++ front/src/applications/operationalStudies/views/Study.tsx
@@ -6,9 +6,9 @@ import { useTranslation } from 'react-i18next';
 import Loader from 'common/Loader';
 import nextId from 'react-id-generator';
 import OptionsSNCF from 'common/BootstrapSNCF/OptionsSNCF';
-import ScenarioCard from 'applications/operationalStudies/components/Study/ScenarioCard';
+import ScenarioCard from 'modules/scenario/components/ScenarioCard';
 import { setFailure } from 'reducers/main';
-import ScenarioCardEmpty from 'applications/operationalStudies/components/Study/ScenarioCardEmpty';
+import ScenarioCardEmpty from 'modules/scenario/components/ScenarioCardEmpty';
 import { VscLink, VscFile, VscFiles } from 'react-icons/vsc';
 import { FaPencilAlt } from 'react-icons/fa';
 import { budgetFormat } from 'utils/numbers';
