+++ front/src/modules/conflict/components/ConflictsList.tsx
