+++ front/src/modules/project/components/DeleteProjectsModal.tsx
