+++ front/src/modules/rollingStock/components/RollingStockSelector/RollingStock2Img.tsx
