+++ front/src/modules/rollingStock/components/RollingStockSelector/RollingStockCard.tsx
