+++ front/src/modules/rollingStock/components/RollingStockSelector/RollingStockCurve.tsx
