+++ front/src/modules/rollingStock/components/RollingStockSelector/RollingStockDarkMode.scss
