+++ front/src/modules/rollingStock/components/RollingStockSelector/RollingStockEmpty.tsx
