+++ front/src/modules/rollingStock/components/RollingStockSelector/RollingStockHelpers.tsx
