+++ front/src/modules/rollingStock/components/RollingStockSelector/RollingStockModal.tsx
