+++ front/src/modules/rollingStock/components/RollingStockSelector/SearchRollingStock.tsx
