+++ front/src/modules/rollingStock/components/RollingStockSelector/WithRollingStockSelector.tsx
