+++ front/src/modules/rollingStock/components/RollingStockSelector/consts/consts.ts
