+++ front/src/modules/rollingStock/components/RollingStockSelector/consts/rollingstockDetailDB.json
