+++ front/src/modules/rollingStock/components/rollingStockEditor/RollingStockEditorFormModal.tsx
