+++ front/src/modules/rollingStock/consts.ts
