+++ front/src/modules/scenario/components/ScenarioCard.tsx
