+++ front/src/modules/scenario/components/ScenarioCardEmpty.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import AddOrEditScenarioModal from 'applications/operationalStudies/components/Scenario/AddOrEditScenarioModal';
+import AddOrEditScenarioModal from 'modules/scenario/components/AddOrEditScenarioModal';
 import { useModal } from 'common/BootstrapSNCF/ModalSNCF';
 import { FaPlus } from 'react-icons/fa';
