+++ front/src/modules/study/components/StudyCard.tsx
