+++ front/src/modules/study/components/StudyCardEmpty.tsx
