+++ front/src/stories/RollingStockSelector.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import RollingStockSelector from 'common/RollingStockSelector/RollingStockSelector';
+import { RollingStockSelector } from 'modules/rollingStock/components/RollingStockSelector';
 import { OSRD_ROLLINGSTOCKSELECTED_SAMPLE_DATA } from 'applications/operationalStudies/components/SimulationResults/SpeedSpaceChart/sampleData';
 import { RollingStock } from 'common/api/osrdEditoastApi';
 import exampleRollingStockImage1 from 'assets/defaultRSImages/example_rolling_stock_image_1.gif';
